+++ app/src/main/java/iths/com/food/Fragments/CategoryFragment.java
@@ -62,7 +62,7 @@ public class CategoryFragment extends Fragment implements OnClickListener{
     }
 
     private void showCategory(String category) {
-        MealFragment newFragment = new MealFragment();
+        MealListFragment newFragment = new MealListFragment();
         Bundle bundle = new Bundle();
         bundle.putString(CHOSEN_CATEGORY, category);
         newFragment.setArguments(bundle);
