--- app/src/main/res/layout/fragment_add_meal.xml
@@ -1,238 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<ScrollView android:layout_height="match_parent"
-    android:layout_width="match_parent"
-    xmlns:android="http://schemas.android.com/apk/res/android">
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/activity_meal"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:paddingBottom="@dimen/activity_vertical_margin"
-    android:paddingLeft="@dimen/activity_horizontal_margin"
-    android:paddingRight="@dimen/activity_horizontal_margin"
-    android:paddingTop="@dimen/activity_vertical_margin"
-    android:orientation="vertical"
-    tools:context="iths.com.food.MealActivity">
-
-    <LinearLayout
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content">
-
-        <TextView
-            android:id="@+id/meal_name_text"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="Testmåltid"
-            android:textSize="32sp"/>
-
-        <Button
-            android:id="@+id/edit_button"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="edit"
-            android:onClick="makeEditable"/>
-
-    </LinearLayout>
-
-    <TextView
-        android:id="@+id/category_text"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Pizza"/>
-
-    <ImageView
-        android:id="@+id/meal_image"
-        android:layout_width="match_parent"
-        android:layout_height="250dp"
-        android:src="@drawable/testmeal" />
-
-    <TextView
-        android:id="@+id/meal_description"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="Asgod mat!"/>
-
-    <TextView
-        android:id="@+id/taste_text"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/taste"/>
-
-    <LinearLayout
-        android:id="@+id/taste_hearts"
-        android:orientation="horizontal"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:src="@mipmap/ic_launcher">
-
-        <ImageView
-            android:id="@+id/heart_taste_1"
-            android:layout_width="20dp"
-            android:layout_height="20dp"
-            android:src="@drawable/empty_heart"
-            />
-
-        <ImageView
-            android:id="@+id/heart_taste_2"
-            android:layout_width="20dp"
-            android:layout_height="20dp"
-            android:src="@drawable/empty_heart"
-            />
-
-        <ImageView
-            android:id="@+id/heart_taste_3"
-            android:layout_width="20dp"
-            android:layout_height="20dp"
-            android:src="@drawable/empty_heart"
-            />
-
-        <ImageView
-            android:id="@+id/heart_taste_4"
-            android:layout_width="20dp"
-            android:layout_height="20dp"
-            android:src="@drawable/empty_heart"
-            />
-
-        <ImageView
-            android:id="@+id/heart_taste_5"
-            android:layout_width="20dp"
-            android:layout_height="20dp"
-            android:src="@drawable/empty_heart"
-            />
-
-        <ImageView
-            android:id="@+id/heart_taste_6"
-            android:layout_width="20dp"
-            android:layout_height="20dp"
-            android:src="@drawable/empty_heart"
-            />
-
-        <ImageView
-            android:id="@+id/heart_taste_7"
-            android:layout_width="20dp"
-            android:layout_height="20dp"
-            android:src="@drawable/empty_heart"
-            />
-
-        <ImageView
-            android:id="@+id/heart_taste_8"
-            android:layout_width="20dp"
-            android:layout_height="20dp"
-            android:src="@drawable/empty_heart"
-            />
-
-        <ImageView
-            android:id="@+id/heart_taste_9"
-            android:layout_width="20dp"
-            android:layout_height="20dp"
-            android:src="@drawable/empty_heart"
-            />
-
-        <ImageView
-            android:id="@+id/heart_taste_10"
-            android:layout_width="20dp"
-            android:layout_height="20dp"
-            android:src="@drawable/empty_heart"
-            />
-
-    </LinearLayout>
-
-    <TextView
-        android:id="@+id/healthy_text"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/healthy"/>
-
-    <LinearLayout
-        android:id="@+id/health_hearts"
-        android:orientation="horizontal"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content">
-
-        <ImageView
-            android:id="@+id/heart_health_1"
-            android:layout_width="20dp"
-            android:layout_height="20dp"
-            android:src="@drawable/empty_heart"
-            />
-
-        <ImageView
-            android:id="@+id/heart_health_2"
-            android:layout_width="20dp"
-            android:layout_height="20dp"
-            android:src="@drawable/empty_heart"
-            />
-
-        <ImageView
-            android:id="@+id/heart_health_3"
-            android:layout_width="20dp"
-            android:layout_height="20dp"
-            android:src="@drawable/empty_heart"
-            />
-
-        <ImageView
-            android:id="@+id/heart_health_4"
-            android:layout_width="20dp"
-            android:layout_height="20dp"
-            android:src="@drawable/empty_heart"
-            />
-
-        <ImageView
-            android:id="@+id/heart_health_5"
-            android:layout_width="20dp"
-            android:layout_height="20dp"
-            android:src="@drawable/empty_heart"
-            />
-
-        <ImageView
-            android:id="@+id/heart_health_6"
-            android:layout_width="20dp"
-            android:layout_height="20dp"
-            android:src="@drawable/empty_heart"
-            />
-
-        <ImageView
-            android:id="@+id/heart_health_7"
-            android:layout_width="20dp"
-            android:layout_height="20dp"
-            android:src="@drawable/empty_heart"
-            />
-
-        <ImageView
-            android:id="@+id/heart_health_8"
-            android:layout_width="20dp"
-            android:layout_height="20dp"
-            android:src="@drawable/empty_heart"
-            />
-
-        <ImageView
-            android:id="@+id/heart_health_9"
-            android:layout_width="20dp"
-            android:layout_height="20dp"
-            android:src="@drawable/empty_heart"
-            />
-
-        <ImageView
-            android:id="@+id/heart_health_10"
-            android:layout_width="20dp"
-            android:layout_height="20dp"
-            android:src="@drawable/empty_heart"
-            />
-
-    </LinearLayout>
-
-    <TextView
-        android:id="@+id/average_grade_text"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Snittbetyg:" />
-
-    <TextView
-        android:id="@+id/average_number"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="0"
-        android:textSize="32sp"/>
-
-</LinearLayout>
-</ScrollView>
